create act table

+++ db/migrate/20190707005712_create_show.rb
@@ -3,7 +3,7 @@ class CreateShow < ActiveRecord::Migration[5.2]
     create_table :shows do |t|
       t.string :name, null: false
       t.datetime :start_time, null: false
-      t.float :duration, null: false
+      t.integer :duration, null: false
       t.text :description
 
       t.timestamps null: false
